Refactor Federation models and adjust naming conventions
Previously, some elements in the Federation models were named inconsistently, and the functionality was spread out across multiple models. This caused some confusion and made the code harder to follow and maintain. This commit helps by refactoring the Models in the Federation domain, specifically: FederationApi.cs, FederationResource.cs, FederationScope.cs, and adding a new file for more clarity: Resource.cs. It also standardizes the naming convention for user claims, changing it from "UserClaims" to "Claims".
