+++ Federation/src/Domain/Models/FederationResource.cs
@@ -5,39 +5,17 @@ using System.Diagnostics;
 namespace Wangkanai.Federation.Models;
 
 /// <summary>
-/// Models the common data Api and identity resources.
+/// Models a user federation resource.
 /// </summary>
-[DebuggerDisplay("{DebuggerDisplay,nq}")]
-public abstract class Resource
-{
-	private string DebuggerDisplay => Name ?? $"{{{typeof(Resource)}}}";
-
-	/// <summary>
-	/// Indicates if this resource is enabled. Defaults to true.
-	/// </summary>
-	public bool Enable { get; set; } = true;
-
-	/// <summary>
-	/// The unique name of the resource.
-	/// </summary>
-	public string Name { get; set; } = default!;
-
-	public string DisplayName     { get; set; }
-	public string Description     { get; set; }
-	public bool   ShowInDiscovery { get; set; } = true;
-
-	public ICollection<string>         UserClaims { get; set; } = new HashSet<string>();
-	public IDictionary<string, string> Properties { get; set; } = new Dictionary<string, string>();
-}
-
+[DebuggerDisplay("{" + nameof(DebuggerDisplay) + ",nq}")]
 public class FederationResource : Resource
 {
-	public FederationResource() { }
-	public FederationResource(string name) : this(name, name) { }
+	private string DebuggerDisplay => Name ?? $"{{{typeof(FederationResource)}}}";
 
-	public FederationResource(string name, string displayName)
-	{
-		Name        = name;
-		DisplayName = displayName;
-	}
+	public FederationResource() { }
+
+	public FederationResource(string name, IEnumerable<string> claims)
+		: this(name, name, claims) { }
+
+	public FederationResource(string name, string displayName, IEnumerable<string> claims) { }
 }
