+++ Federation/src/Domain/Models/FederationScope.cs
@@ -32,6 +32,6 @@ public class FederationScope : FederationResource
 			return;
 
 		foreach (var claim in claims)
-			UserClaims.Add(claim);
+			Claims.Add(claim);
 	}
 }
